Class

+++ Class/Class.Console/GenArg.cs
@@ -2,6 +2,6 @@ namespace Class.Console;
 
 public class GenArg : Any
 {
-    public virtual int Index { get; set; }
+    public virtual long Index { get; set; }
     public virtual Data Data { get; set; }
 }
